Refactored code slightly to support documentation generation, moved enum into separate file
