+++ Commands/Enums/AlertFilter.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SharePointPnP.PowerShell.Commands.Enums
+{
+    public enum AlertFilter
+    {
+        AnythingChanges = 0,
+        SomeoneElseChangesAnItem = 1,
+        SomeoneElseChangesItemCreatedByMe = 2,
+        SomeoneElseChangesItemLastModifiedByMe = 3
+    }
+}
